Refactor and validate rust codebase (#116)
* Refactor time_reversal and visualization modules into focused submodules


* Refactor therapy module into focused submodules with improved architecture


* Fix trait method ambiguity and add grid parameter to array access methods


* Fix compilation errors and update version to 2.58.0


* Prepare v3.0.0: Production build with test success and solver refactoring


* Refactor unified solver into modular acoustic wave solvers


* Refactor performance module into focused, modular optimization components


* Refactor AMR module: Modularize, simplify, and improve core implementations


* Refactor constants, remove old constants module, update imports


* Refactor GPU module, add compute kernels and GPU acceleration support


* Add PAM sensor module and update constants with new parameters


* Refactor passive acoustic mapping with modular architecture


* Implement GPU support, refactor PAM, and update project version to 4.1.0


* Implement zero stubs, robust Capon beamformer, and GPU pipeline


* Refactor field analysis, update beamforming, and bump version to 4.3.0


---------

+++ tests/literature_validation.rs
@@ -118,8 +118,8 @@ fn test_elastic_wave_velocities() {
     let lambda = rho * vp * vp - 2.0 * mu;
 
     // Verify relationships
-    let vp_calc = ((lambda + 2.0 * mu) / rho).sqrt();
-    let vs_calc = (mu / rho).sqrt();
+    let vp_calc = ((lambda + 2.0 * mu) / rho as f64).sqrt();
+    let vs_calc = (mu / rho as f64).sqrt();
 
     assert!((vp_calc - vp).abs() < 1.0, "P-wave velocity mismatch");
     assert!((vs_calc - vs).abs() < 1.0, "S-wave velocity mismatch");
